added checks to ignore git directories

+++ ModManager.StarCraft.Base/PathUtils.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Text.RegularExpressions;
 using Microsoft.Win32;
 using ModManager.StarCraft.Base.Enums;
 
@@ -32,6 +33,10 @@ namespace ModManager.StarCraft.Base
             //Now Create all of the directories
             foreach (string dirPath in Directory.GetDirectories(sourcePath, "*", SearchOption.AllDirectories))
             {
+                if (IsGitPath(dirPath))
+                {
+                    continue;
+                }
                 Directory.CreateDirectory(dirPath.Replace(sourcePath, targetPath));
                 Console.WriteLine($"Created Dir at '{dirPath.Replace(sourcePath, targetPath)}'");
             }
@@ -41,6 +46,11 @@ namespace ModManager.StarCraft.Base
             {
                 File.Copy(newPath, newPath.Replace(sourcePath, targetPath), true);
             }
+        }
+
+        public static bool IsGitPath(string path)
+        {
+            return Regex.IsMatch(path, ".*\\.git.*");
         }
 
         public static bool TryClearDirectory(string directory, Action<string> reportError)
@@ -70,7 +80,7 @@ namespace ModManager.StarCraft.Base
             //TODO: Fix the evo missions
             foreach (string subdir in Directory.GetDirectories(directory, "*", SearchOption.TopDirectoryOnly))
             {
-                if (!(subdir.Contains(@"Campaign\swarm") || subdir.Contains(@"Campaign\void") || subdir.Contains(@"Campaign\nova")))
+                if (!IsGitPath(subdir) && !(subdir.Contains(@"Campaign\swarm") || subdir.Contains(@"Campaign\void") || subdir.Contains(@"Campaign\nova")))
                 {
                     Directory.Delete(subdir, true);
                 }
